refactor: change the name of utils folders to libs to be more declarative with the modules design that we doing on the project
