+++ src/components/Form/SignInModal/SignIn/index.tsx
@@ -9,10 +9,10 @@ import {
 } from '@/components/Form/SignInModal/styles'
 import { isSome, none, chain, getLeft, fromNullable } from 'fp-ts/Option'
 import { pipe } from 'fp-ts/lib/function'
-import { f } from '@/utils/expression'
+import { f } from '@/libs/expression'
 import { useMutation } from '@tanstack/react-query'
 import { Either } from 'fp-ts/lib/Either'
-import { DefaultError } from '@/utils/errors'
+import { DefaultError } from '@/libs/errors'
 import { SignInInput } from '@/types/user'
 import { signInMutation, type SignInResponseOutput } from './signInMutation'
 import * as Yup from 'yup'
