+++ src/components/Form/SignInModal/SignUp/index.tsx
@@ -7,8 +7,8 @@ import {
 } from '@/components/Form/SignInModal/styles'
 import { useAuth } from '@/context/auth'
 import { SignInInput } from '@/types/user'
-import { DefaultError } from '@/utils/errors'
-import { f } from '@/utils/expression'
+import { DefaultError } from '@/libs/errors'
+import { f } from '@/libs/expression'
 import { useMutation } from '@tanstack/react-query'
 import { useFormik } from 'formik'
 import { Either } from 'fp-ts/Either'
